requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
async-lru==1.0.2
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
colorama==0.4.4
idna==3.3
importlib-metadata==4.8.1
iniconfig==1.1.1
packaging==21.2
pluggy==1.0.0
py==1.11.0
pyaes==1.6.1
pyparsing==2.4.7
Pyrogram==1.2.9
PySocks==1.7.1
pytest==6.2.5
pytest-html==3.1.1
pytest-metadata==1.11.0
requests==2.26.0
selenium==3.141.0
six==1.16.0
TgCrypto==1.2.2
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
zipp==3.6.0
